requirements.txt
spotipy==2.22.0
python-dotenv==0.21.0
et-xmlfile==1.1.0
